+++ examples/poll/src/components/WidgetWrapper/index.tsx
@@ -1,13 +1,13 @@
 import { Box } from 'grommet';
 import { Drag } from 'grommet-icons';
-import { PropsWithChildren } from 'react';
+import { SelectWidgetType } from ':/data/getLocalizedWidgetDefinitions';
 
-export default function WidgetWrapper({ children }: PropsWithChildren<{}>) {
+export default function WidgetWrapper({ text }: SelectWidgetType) {
   return (
     <Box align="center" direction="row" gap="5px">
       <div className="form-toolbar-item">
         <Drag color="#7D4CDB77" size="medium" />
-        {children}
+        {text}
       </div>
     </Box>
   );